requirements.txt
selenium==3.8.0
Pillow==4.3.0
unicode_slugify==0.1.3
